Category deletion and comment storage
FOR NEXT TIME:
- See how to reload the comment form
- Do the edit part

-----------
- I can only delete categories if they are not attached to anything
- I can create new comments on a ticket page

+++ app/Http/Controllers/CategoryController.php
@@ -67,7 +67,11 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
-        $category->tickets()->detach();
-        $category->delete();
+        if ($category->tickets()->get()->count() === 0) {
+            $category->tickets()->detach();
+            $category->delete();
+            //$categories = Category::all();
+
+        } else return response()->json(["message" => "This category is attached to tickets, you cannot delete it!!"], 400);
     }
 }
